Login Handels Driver Or Standard User
Modified the login class and the server side PHP script so that it will
tell if a user is considered a driver or a standard app user, as Wei
discussed in class.
WEI : I added two comments where I think you can add the calls to your
driver or user map activities.
A test driver account in the MySQL server is user: 'a' pass: 'a'